feat: properly handle redrawing of deorations
This commit now only handles the refresh of decorations when switching
to another code review CSV, but also fixes an issue with duplicating
speech bubbles. The issue most likely has been that the wrong mechanis
has been implemented to dispose stale decorations. In the documention
of editor.setDecorations it is stated that the decorations are cleared
for the given type and editor if the list of options/ranges is empty.
Caveat: The given type must stay the same. And this was probably the
root of the bug: New types were generated every time the decorations
were redrawn. This commit fixes that and initalizes them only once.