feat: export as markdown (#133)
* fix: show error message when default template not found

When the default template is not found, we should notify the user so
they don't think the template is still processing.

When unable to locate the file, fs.readFileSync throws an error. Catch
the error and show a window error message.

* feat: add support for exporting as markdown

Add additional export file format for Markdown export. Allow users to
select a custom export file in the settings.

* feat: add default markdown template

Create a standard template which will be used by default for markdown
exports. This is almost identical to the HTML default template.

* feat: copy markdown template to dist

* feat: open markdown file after export

After successfully exporting the markdown file, open it in the side
panel.

* fix: correct broken anchor reference

Fix the link referenced to point to the header.

* docs: add new markdown setting and command to docs